+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 
 - Support for nodeinfo 2.1 and 2.2.
 - ActorStorage method for checking active users.
+- nodeinfo properties for monthly active and half-yearly active users.
 
 ### Changed